% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_source.R
\name{fit_pclm}
\alias{fit_pclm}
\title{Fit univariate pclm}
\usage{
fit_pclm(dta, breaks, offset = NULL, lambda, deg = 2, show = FALSE,
  ci.level = 0.05, iter = 50, tol = 1e-06)
}
\description{
Fit univariate pclm
}
\keyword{internal}
