% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmddx}
\alias{hmddx}
\title{Data in the package (for testing purposes)}
\format{An object of class \code{data.frame} with 501498 rows and 6 columns.}
\source{
\href{http://www.mortality.org}{Human Mortality Database}
}
\usage{
hmddx
}
\description{
Dataset containing death counts (Dx) from Human Mortality Database.
}
\keyword{datasets}
