% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmdEx}
\alias{hmdEx}
\title{Data in the package (for testing purposes)}
\format{An object of class \code{data.frame} with 29304 rows and 6 columns.}
\source{
\href{http://www.mortality.org}{Human Mortality Database}
}
\usage{
hmdEx
}
\description{
Dataset containing exposures (Ex) for Sweden from Human Mortality Database.
}
\keyword{datasets}
