% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm-package.R
\docType{package}
\name{pclm-package}
\alias{pclm-package}
\title{Penalised Composite Link Model (PCLM) tools in R}
\description{
Simple tool for PCLM using iteratively re-weighted least squares.
}
\details{
Essentially provides a wrapper around the (corrected) code from

Silvia Rizzi, Jutta Gampe, and Paul H. C. Eilers, 2015, Efficient Estimation
of Smooth Distributions From Coarsely Grouped Data Am. J. Epidemiol. first
published online June 16, 2015 doi.10.1093/aje/kwv020

which finds the optimal penalty in the sense that it minimises the Akaike
information criteria.

This code is licensed under the MIT license
\url{https://www.r-project.org/Licenses/MIT}, and you may use this package
strictly under those terms
}
\author{
Stephen Wade \email{stephematician@gmail.com}
}
