% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_mx.R
\name{pclm_mx}
\alias{pclm_mx}
\title{Estimate death rates using PCLM}
\usage{
pclm_mx(dta, offset, breaks, lambda = NULL, show = FALSE, ci.level = 0.05,
  objective.fun = "AIC", opt.interval = c(0, 10^5))
}
\arguments{
\item{dta}{Vector of observed counts}

\item{offset}{Offset term.}

\item{breaks}{Breaks in the data set. For example: if we have 3 bins 
\code{[0-5), [5-10) and [10-15]} the breaks will be defined as a vector 
\code{c(0, 5, 10, 16)}}

\item{lambda}{Smoothing parameter. If \code{lambda = NULL} (default) an 
algorithm will find the optimal value.}

\item{show}{Logical value. Indicates whether iteration details should be shown or not.
Default: \code{FALSE}.}

\item{ci.level}{Level of significance for computing confidence intervals. 
Default: \code{0.05}.}

\item{objective.fun}{Objective function used in optimisation process.
Choose between \code{AIC} or \code{BIC}. Default: \code{AIC}}

\item{opt.interval}{Interval to be used in one-dimensional optimisation of lambda. 
Default: \code{c(0, 10^5)}.}
}
\value{
The function will return 3 pclm objects: one for dta data, 
one for data provided in the offset argument, and one for the death rates.
The first two pclm objects can be obtained by fitting two separate \code{\link{pclm}}s too, 
using the two data sets (dta and offset).
}
\description{
Estimate 1-year age-specific death rates from aggregated data 
using the univariate penalized composite link model.
}
\examples{
library(pclm)

# Select data
year = 2014
sex  = "Total"

exposure <- hmdEx[hmdEx$Year == year, ]
deaths <- hmdDx[hmdDx$Year == year, ]

# Aggregate them artificially in 5-years age classes with 85+
# add last bin with 0 counts to complete the histogram
breaks = c(seq(0, 85, by = 5), 115, 130)
exposure$Groups_Counts = deaths$Groups_Counts <- c(rep(1:17, each = 5), rep(18, 26))

aggredate_Dx <- aggregate(deaths$Total, FUN = "sum",
                          by = list(deaths$Groups_Counts))$x
aggredate_Dx <- c(aggredate_Dx, 0)

aggredate_Ex <- aggregate(exposure$Total, FUN = "sum",
                          by = list(exposure$Groups_Counts))$x
aggredate_Ex <- c(aggredate_Ex, 0)

# Fit the models ---------------------
model <- pclm_mx(dta = aggredate_Dx, offset = aggredate_Ex, breaks)

M1 = model$pclm_dta
M2 = model$pclm_offset
M3 = model$pclm_mx

# Plot1 - Observed and estimated death distribution
par(mfrow = c(1, 3))
plot(M1)
lines(0:110, deaths$Total, type = "o", col = "blue")

# Plot 2 - Observed and estimated distribution of population exposed to risk
plot(M2)
lines(0:110, exposure$Total, type = "o", col = "blue")

# Plot 3 - Age-specific death rate estimation
plot(M3)
mx_observed  <- deaths[, 'Total']/exposure[, 'Total']
points(mx_observed)

}
\seealso{
\code{\link{pclm}}
}
