% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclmIRLS.R
\name{pclmIRLS}
\alias{pclmIRLS}
\title{Iteratively re-weighted least squares for Penalised Composite Link Models}
\usage{
pclmIRLS(n, C, X = diag(nrow = ncol(C)), lambda = 1, deg = 2)
}
\arguments{
\item{n}{a numeric vector containing grouped count data}

\item{C}{a matrix for the mapping of bins to groups in the model}

\item{X}{a matrix to link a parameterisation of the probability
distribution for the binned data}

\item{lambda}{a scalar smoothness penalty}

\item{deg}{the degree of the smoothness operator}
}
\value{
a list with the following names components
\itemize{
    \item{'mu'}{The (fitted) expected count in each group}
    \item{'gamma'}{The (fitted) expected count in each bin}
    \item{'trace'}{The effective dimension of the (fitted) model}
    \item{'dev'}{The deviance of gamma given mu}
    \item{'aic'}{The Aikake Information Criterion of the fitted model}
    \item{'converged'}{Convergence status of IRLS procedure}
}
}
\description{
Applies IRLS to a find the maximum value of a penalised likelihood expression
representing the likelihood of a composite link model.
}
\details{
Silvia Rizzi, Jutta Gampe, and Paul H. C. Eilers, 2015, Efficient Estimation
of Smooth Distributions From Coarsely Grouped Data Am. J. Epidemiol. first
published online June 16, 2015 doi.10.1093/aje/kwv020
}
