% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpclmfit.R
\name{rpclmfit}
\alias{rpclmfit}
\title{Randomly draw samples from fitted PCLM}
\usage{
rpclmfit(D, pclm_fit, n_sample = 1L)
}
\arguments{
\item{D}{A two-column data.frame of integers, each row represents an
observation of an interval of bins, columns correspond to lower and
upper bins of the interval (inclusive)}

\item{pclm_fit}{A list as returned by pclmfit}

\item{n_sample}{Number of samples to draw for each observation}
}
\value{
Matrix of drawn bins
}
\description{
Scales reasonably well (linear?) with
\itemize{
  \item number of observations
  \item number of unique (observed) intervals
  \item number of samples per interval
}
}
\details{
Takes about 1 second on a 2016 desktop, Windows 7, for
\itemize{
  \item 100000 observations
  \item 2500 unique intervals
  \item 5 samples per interval
}

Cannot handle PCLMs with more than \code{sqrt(.Machine$max.integer)} number
of bins due to a hashing mechanism used
}
