% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.pclm}
\alias{summary.pclm}
\title{Summary of the fitted PCLM object}
\usage{
summary.pclm(object, ...)
}
\arguments{
\item{object}{Fitted PCLM object.}
}
\description{
\emph{bold{Generic function}}
}
\seealso{
\code{\link{plot.pclm}}
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}
