% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_source.R
\name{objective_fun}
\alias{objective_fun}
\title{Objective function (to be optimised)}
\usage{
objective_fun(x, y, breaks, what = "aic")
}
\description{
Objective function (to be optimised)
}
\keyword{internal}
