% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_source.R
\name{pclm}
\alias{pclm}
\title{Fit Univariate Penalized Composite Link Model}
\usage{
pclm(y, breaks, lambda = NULL, deg = 2, show = FALSE)
}
\arguments{
\item{y}{Vector of observed counts}

\item{breaks}{Breaks}

\item{lambda}{Smoothing parameter. If \code{lambda = NULL} (default) an 
algorithm will find the optimal value.}

\item{deg}{Order of differences of the components of b}

\item{show}{a logical value. Indicates whether iteration details should be shown}
}
\value{
A \code{pclm} object
}
\description{
Fit Univariate Penalized Composite Link Model
}
\examples{
 
library(pclm)

# Select data
country = "SWE"
year    = 2014
sex     = "Total"

deaths <- hmddx[hmddx$country == country & hmddx$Year == year, ]

# Aggregate them artificially in 5-years age classes with 85+ 
# add last bin with 0 counts to complete the histogram
breaks = c(seq(5,85,5),115,130)
deaths$Groups_Counts <- c(rep(1:17, each = 5), rep(18, 26))
aggredate_dx <- aggregate(deaths$Total, FUN = "sum",
                          by = list(deaths$Groups_Counts))$x
aggredate_dx <-  c(aggredate_dx, 0)


# Fit model
mod = pclm(y = aggredate_dx, breaks)

# Generic plot
plot(mod)
# Add real data on top
lines(0:110, deaths$Total, type = "o", col = "blue") 

}
