% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclmfit.R
\name{pclmfit}
\alias{pclmfit}
\title{Penalised Composite Link Model by minimisation of AIC}
\usage{
pclmfit(n, x, w = rep(1, length(x)), use_spline = F, k = 4)
}
\arguments{
\item{n}{a numeric vector of counts in each group}

\item{x}{a numeric vector of the index into n for each bin}

\item{w}{a numeric vector of (within group) weights for each bin}

\item{use_spline}{a boolean indicating whether or not to use a spline for the
basis of the expected count in each bin (default false)}

\item{k}{the number of knots to use in the spline describing the expected
count in each bin (default 4)}
}
\value{
a list with the following named components
\itemize{
    \item{'mu'}{The (fitted) expected count in each group}
    \item{'gamma'}{The (fitted) expected count in each bin}
    \item{'trace'}{The effective dimension of the (fitted) model}
    \item{'dev'}{The deviance of gamma given mu}
    \item{'aic'}{The Aikake Information Criterion of the fitted model}
    \item{'converged'}{Convergence status of IRLS procedure}
}
}
\description{
No description here yet
}
\details{
Description
}
