% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclmfit.R
\name{pclmfit}
\alias{pclmfit}
\title{Penalised Composite Link Model by minimisation of AIC}
\usage{
pclmfit(n, x, w = rep(1, length(x)), use_spline = F, k = 4)
}
\arguments{
\item{n}{a numeric vector of counts in each group}

\item{x}{a numeric vector of the index into n for each bin}

\item{w}{a numeric vector of (within group) weights for each bin}

\item{use_spline}{a boolean indicating whether or not to use a spline for the
basis of the expected count in each bin (default false)}

\item{k}{the number of knots to use in the spline describing the expected
count in each bin (default 4)}
}
\value{
a list with the following named components
\itemize{
    \item{'mu'}{The (fitted) expected count in each group}
    \item{'gamma'}{The (fitted) expected count in each bin}
    \item{'trace'}{The effective dimension of the (fitted) model}
    \item{'dev'}{The deviance of gamma given mu}
    \item{'aic'}{The Aikake Information Criterion of the fitted model}
    \item{'converged'}{Convergence status of IRLS procedure}
    \item{'lambda'}{Value of smoothing parameter that minimised AIC}
}
}
\description{
Fits a penalised composite link model (PCLM) to grouped count data. This
provides an estimate of the distribution of objects in a specified number
of bins within each group (typically equi-spaced).
}
\details{
The PCLM is given by minimising the Akaike Information Criterion (to specify
the smoothing parameter) as outlined in the paper:

Silvia Rizzi, Jutta Gampe, and Paul H. C. Eilers, 2015, Efficient estimation
of smooth distributions from coarsely grouped data, \emph{Am. J. Epidemiol.,
182}(2), pp. 138-147, doi.10.1093/aje/kwv020

Allows for a within-group weighting in each bin, so that additional
(relative) information about the rates within a group can be incorporated.
For example, given two bins within a group, it might be known that one bin
is twice as likely as the other (aside: this is not the same as knowing that
one bin is twice as 'wide' as the other).

If there are a large number of bins, it can be useful to model the rates (in
each bin) using splines. This is possible by specifying \code{use_spline=T}
and selecting a number of knots using the \code{k} argument (defaults to
four). The knot placement is given by the default behaviour of
\code{\link[splines]{bs}}. For further discussion see:

Paul H. C. Eilers and Brian D. Marx, 1996, Flexible smoothing with B-splines
and penalties, \emph{Statistical Science, 11}(2), pp. 89-102.
}
